requirements.txt
Django==3.1
django-allauth==0.42.0
django-crispy-forms==1.9.2
django-debug-toolbar==2.2
django-environ==0.4.5
django-extensions==3.0.3
gunicorn==20.0.4
ipython==7.17.0
psycopg2-binary==2.8.5
ptyprocess==0.6.0
pytest==6.0.1
pytest-cov==2.10.0
pytest-django==3.9.0